Reorganização para arquitetura hexagonal
- Exclusão da pasta Web
- Criação da pasta Adapters
- Ajuste na nomenclatura dos arquivos
