+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node dist/server.js",
-    "dev": "tsx watch src/server.ts",
+    "dev": "tsx watch src/infraestructure/server.ts",
     "build": "tsup src",
     "test": "vitest --run --coverage --passWithNoTests",
     "test:watch": "vitest --watch",
