+++ src/adapters/index.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import { orderRoutes } from './order';
+import { orderRoutes } from './order/controller/order';
 
 const routes = express.Router();
 routes.use('/orders', orderRoutes);
