+++ src/infrastructure/app.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import helmet from 'helmet';
-import { routes } from './web/controllers';
+import { routes } from '../adapters';
 
 const app = express();
 app.use(helmet());
